agent/requirements.txt
wxPython==4.2.1
Pillow==10.4.0
psutil==6.0.0
websocket-client==1.8.0
pywin32==306
pyautogui==0.9.54
GPUtil==1.4.0